sprint 3 finished

+++ server.py
@@ -1,6 +1,6 @@
 from flask import Flask, render_template, request, redirect, url_for
 from flask import session, escape
-import util, database_manager, hashing_utility
+import util, database_manager, hashing_utility, session_controller
 import os
 from datetime import datetime, timedelta
 
@@ -8,7 +8,7 @@ app = Flask(__name__)
 
 ###################### SESSION SETTINGS ######################
 app.secret_key = b'\xa0\xd7\xe2\x8er\xd6\xd2\xd8\x8b\xce\xb3T'
-app.permanent_session_lifetime = timedelta(minutes=5)
+app.permanent_session_lifetime = timedelta(minutes=10)
 ###############################################################
 
 
@@ -20,9 +20,14 @@ def index():
     if 'username' in session:
 
         username = session['username']
+        userID = database_manager.getUserIDbyUsername(username)
+        ###### SESSION REFRESH #######
+        session_controller.session_refresh(session)
+        ##############################
+
 
         return render_template('index.html', questions=last_five_questions,
-                               page_title='AskMate', logged=True, username=username )
+                               page_title='AskMate', username=username, userID=userID)
 
     return render_template('index.html', questions = last_five_questions,
                            page_title = 'AskMate')
@@ -30,6 +35,7 @@ def index():
 @app.route('/list')
 def list():
 
+    session_controller.session_refresh(session)
     question_file = database_manager.import_questions_for_list()
     titles = ['ID','Submission time','Title',]
     sort = request.args.get('sort')
@@ -56,24 +62,28 @@ def list():
 @app.route('/add_question', methods=['GET','POST'])
 def add():
 
-    if request.method == "POST":
+    userID = session_controller.session_refresh(session)
 
+    if userID:
 
-        title = request.form['title']
-        message = request.form['message']
-        id = database_manager.highest_id()
-        view = 0
-        vote = 0
-        image = None
-        timestamp = datetime.now().replace(microsecond=0)
+        if request.method == "POST":
 
-        values = (id,timestamp,view,vote,title,message,image)
-        database_manager.add_question(values)
+            title = request.form['title']
+            message = request.form['message']
+            id = database_manager.highest_id()
+            view = 0
+            vote = 0
+            image = None
+            timestamp = datetime.now().replace(microsecond=0)
 
+            values = (id,timestamp,view,vote,title,message,image,userID)
+            database_manager.add_question(values)
 
-        return redirect (url_for('question',id=id))
+            return redirect (url_for('question',id=id))
 
-    return render_template('add_question.html', page_title = 'AskMate - Add question')
+        return render_template('add_question.html', page_title = 'AskMate - Add question')
+
+    return redirect('/login')
 
 @app.route('/question/<id>', methods=['GET','POST'])
 def question(id):
@@ -83,18 +93,36 @@ def question(id):
     except ValueError:
         return redirect('/list')
 
+    userID = session_controller.session_refresh(session)
     question = database_manager.show_question(id)
     answers = database_manager.show_answers(id)
     page_title = 'Question ' + str(id)
 
     if question:
 
+        if userID:
+
+            questionOwner = database_manager.getQuestionOwner(id)
+
+            if userID == questionOwner:
+
+                database_manager.view_counter(question_id=id)
+
+                return render_template('question.html', id=id,question=question,answers=answers,
+                                       page_title=page_title, questionAccess=True, logged=True,
+                                       userID=userID)
+
+            database_manager.view_counter(question_id=id)
+
+            return render_template('question.html', id=id,question=question, answers=answers,
+                               page_title=page_title, logged=True, userID=userID)
+
         database_manager.view_counter(question_id=id)
 
-        return render_template('question.html', id=id,question=question, answers=answers, page_title=page_title)
+        return render_template('question.html', id=id, question=question, answers=answers,
+                               page_title=page_title)
 
-
-    return redirect('/list')
+    return redirect(url_for('index'))
 
 @app.route('/question/<id>/new-answer', methods=['GET','POST'])
 def add_answer(id):
@@ -106,31 +134,37 @@ def add_answer(id):
 
     if database_manager.is_id_in_database(id):
 
-        if request.method == 'POST':
+        userID = session_controller.session_refresh(session)
 
-            answer_id = database_manager.highest_answer_id()
-            timestamp = datetime.now().replace(microsecond=0)
-            vote_number = 0
-            question_id = id
-            message = request.form['message']
-            image = None
+        if userID:
 
-            values = (answer_id,timestamp,vote_number,question_id,message,image)
-            database_manager.add_answer(values)
+            if request.method == 'POST':
+
+                answer_id = database_manager.highest_answer_id()
+                timestamp = datetime.now().replace(microsecond=0)
+                vote_number = 0
+                question_id = id
+                message = request.form['message']
+                image = None
+
+                values = (answer_id,timestamp,vote_number,question_id,message,image,userID)
+                database_manager.add_answer(values)
 
 
-            return redirect(url_for('question', id=question_id))
+                return redirect(url_for('question', id=question_id))
 
-        return render_template('new_answer.html', id=id)
+            return render_template('new_answer.html', id=id)
 
+        else:
+            return redirect(url_for('login'))
 
-    return redirect('/')
+    return redirect(url_for('index'))
 
 @app.route('/answer/<id>', methods=['GET','POST'])
 def answer(id):
 
     page_title = 'Answer ' + str(id)
-    index = {'ANSWER_ID':0,'TIMESTAMP':1,'VOTE':2,'QUESTION_ID':3,'MESSAGE':4,'IMAGE':5}
+    index = {'ANSWER_ID':0,'TIMESTAMP':1,'VOTE':2,'QUESTION_ID':3,'MESSAGE':4,'IMAGE':5,'USER_ID':6}
 
     try:
         id = int(id)
@@ -141,7 +175,20 @@ def answer(id):
 
     if answer:
 
-        answer = answer[0]
+        userID = session_controller.session_refresh(session)
+
+        if userID:
+
+            answerOwner = database_manager.getAnswerOwner(id)
+
+            if userID == answerOwner:
+
+                return render_template('answer.html', id=id, answer=answer, page_title=page_title,
+                                        index=index,logged=True,answerAccess=True)
+
+            return render_template('answer.html',id=id,answer=answer,page_title=page_title,
+                                    index=index,logged=True)
+
 
         return render_template('answer.html',id=id, answer=answer, page_title=page_title, index=index)
 
@@ -157,15 +204,20 @@ def delete_question(id):
 
     if database_manager.is_id_in_database(id):
 
-        if request.method == 'POST':
+        userID = session_controller.session_refresh(session)
+        questionOwner = database_manager.getQuestionOwner(id)
 
-            database_manager.delete_record_from_question(id)
+        if userID == questionOwner:
 
-            return redirect('/list')
+            if request.method == 'POST':
 
-        return render_template('delete_question.html', id=id)
+                database_manager.delete_record_from_question(id)
 
-    return redirect('list')
+                return redirect('/list')
+
+            return render_template('delete_question.html', id=id)
+
+    return redirect(url_for('list'))
 
 @app.route('/question/<id>/vote', methods=['GET','POST'])
 def vote(id):
@@ -175,23 +227,25 @@ def vote(id):
     except ValueError:
         return redirect('/list')
 
-    question = database_manager.is_id_in_database(id=id)
+    if database_manager.is_id_in_database(id=id):
 
-    if question:
+        userID = session_controller.session_refresh(session)
 
-        if request.method == 'POST':
+        if userID:
 
-            if request.form['button'] == 'up':
+            if request.method == 'POST':
 
-                database_manager.vote_up(question_id=id)
-                return redirect (url_for('question',id=id))
+                if request.form['button'] == 'up':
 
-            elif request.form['button'] == 'down':
+                    database_manager.vote_up(question_id=id)
+                    return redirect (url_for('question',id=id))
 
-                database_manager.vote_down(question_id=id)
-                return redirect(url_for('question',id=id))
+                elif request.form['button'] == 'down':
 
-        return render_template('vote.html', id=id)
+                    database_manager.vote_down(question_id=id)
+                    return redirect(url_for('question',id=id))
+
+            return render_template('vote.html', id=id)
 
     return redirect('/')
 
@@ -205,17 +259,19 @@ def vote_answer(id):
 
     answer = database_manager.show_answer_by_id(id)
     page_title = 'Vote Answer ' + str(id)
+    userID = session_controller.session_refresh(session)
 
     if answer:
-        if request.method == 'POST':
-            if request.form['button'] == 'up':
-                database_manager.answer_vote_up(id)
-                return redirect(url_for('answer', id=id))
-            else:
-                database_manager.answer_vote_down(id)
-                return redirect(url_for('answer', id=id))
+        if userID:
+            if request.method == 'POST':
+                if request.form['button'] == 'up':
+                    database_manager.answer_vote_up(id)
+                    return redirect(url_for('answer', id=id))
+                else:
+                    database_manager.answer_vote_down(id)
+                    return redirect(url_for('answer', id=id))
 
-        return render_template('vote_answer.html', id=id, page_title=page_title)
+            return render_template('vote_answer.html', id=id, page_title=page_title)
 
     return redirect('/')
 
@@ -230,11 +286,17 @@ def delete_answer(id):
     answer = database_manager.show_answer_by_id(id)
 
     if answer:
-        if request.method == 'POST':
-            database_manager.delete_answer(id)
-            return redirect('/list')
 
-        return render_template('delete_answer.html', id=id)
+        userID = session_controller.session_refresh(session)
+        answerOwner = database_manager.getAnswerOwner(id)
+
+        if userID == answerOwner:
+
+            if request.method == 'POST':
+                database_manager.delete_answer(id)
+                return redirect('/list')
+
+            return render_template('delete_answer.html', id=id)
 
     return redirect('/')
 
@@ -304,7 +366,7 @@ def login():
 
             if hashing_utility.verify_password(password,hashedPassword):
 
-                session.permament = True
+                session.permanent = True
                 session['username'] = username
 
                 return redirect(url_for('index'))
@@ -336,17 +398,26 @@ def user(id):
     except ValueError:
         return redirect('/')
 
+    USERNAME_INDEX = 2
     userData = database_manager.getUserById(id)
+    userName = userData[USERNAME_INDEX]
+    userQuestions = database_manager.getUserQuestions(id)
+    userAnswers = database_manager.getUserAnswers(id)
     page_title = 'AskMate - User ' + str(id)
+    questionTitles = ['ID','Title','Message','Submission Time','Views','Rating']
+    answerTitles = ['ID','Message','Submission Time','Rating']
+
 
     if userData:
 
-        return render_template('user.html', userData=userData, page_title = page_title)
+        return render_template('user.html', userName=userName, userQuestions=userQuestions,
+                               page_title = page_title, questionTitles=questionTitles,
+                               userAnswers=userAnswers,answerTitles=answerTitles)
 
     return redirect('/')
 
 if __name__ == "__main__":
     app.run(
-        debug=True, # Allow verbose error reports
+        debug=False, # Allow verbose error reports
         port=5000 # Set custom port
     )
